# api/docs/swagger.yaml
basePath: /api
definitions:
  audit_log.AuditLog:
    properties:
      action:
        description: action and reason json string
        type: string
      changes:
        description: Full JSON or JSON diff
        type: string
      created_at:
        type: string
      id:
        type: integer
      model:
        description: e.g., "OcservUser"
        type: string
      model_id:
        description: e.g., UID or primary key
        type: string
      user_uid:
        description: ULID or user identifier
        type: string
      username:
        type: string
    type: object
...
  log.UsersLogsResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/audit_log.AuditLog'
        type: array
    required:
    - meta
    type: object
  middlewares.PermissionDenied:
    properties:
      error:
        type: string
    type: object
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  models.ServerVersion:
    properties:
      occtl_version:
        type: string
      version:
        type: string
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /logs/audit:
    get:
      consumes:
      - application/json
      description: List of logs
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Search User by UID
        in: query
        name: uid
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/log.UsersLogsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: List of logs
      tags:
      - Logs
  /logs/users:
    get:
      consumes:
      - application/json
      description: List of Users logs on self user model
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/log.UsersLogsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: List of Users logs on self user model
      tags:
      - Logs(Users)
  /occtl/commands:
    get:
      consumes:
      - application/json
      description: Occtl Commands
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Command Action ID (1 to 15)
        in: query
        name: action
        required: true
        type: integer
      - description: Optional parameter depending on command
        in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Occtl Commands
      tags:
      - OCCTL
  /occtl/server_info:
    get:
      consumes:
      - application/json
      description: Server information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServerVersion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Server information
      tags:
      - OCCTL
...
swagger: "2.0"